+++ database/migrations/2024_04_16_005300_create_distribusi_surats_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('idSuratMasuk');
             $table->foreign('idSuratMasuk')->references('id')->on('surat_masuk');
-            $table->unsignedBigInteger('idTujuanDisposisi');
+            $table->unsignedInteger('idTujuanDisposisi');
             $table->foreign('idTujuanDisposisi')->references('id')->on('users');
-            $table->unsignedBigInteger('idPengirimDisposisi');
+            $table->unsignedInteger('idPengirimDisposisi');
             $table->foreign('idPengirimDisposisi')->references('id')->on('users');
             $table->date('tanggalDiteruskan');
             $table->string('status');
